hue: add hue_scene and cleanup
- rename tzLocal.effect -> tzLocal.hue_scene
- update exposes to take disableHueScenes option
- set disableHueScenes: false for LTG002, LCG002, LCA008 (bulbs I own and was able to test)
- move hue specific converters to devices/philips.js
